refactored and improved
